Two little fixes...

+++ ShaderGraphNode_FadeOverDistance.cs
@@ -23,6 +23,7 @@ namespace Jick
 		/// <param name="Position">The position to measure distance from (usually set this to the camera position).</param>
 		/// <param name="Distances">The distance values to fade at (X:The near fade start distance, Y:The near fade end distance, Z:The far fade start distance, W:The far fade end distance).</param>
 		/// <param name="Alpha">Whether to return an alpha value (for alpha fading) or a dithered value (0 or 1).</param>
+		/// <param name="Jitter">Whether to apply jitter (small random variation) to the dither in order to mask repeated/uniform areas.</param>
 		/// <param name="WorldSpacePosition">The position of the current pixel in world space (usually don't need to set this).</param>
 		/// <param name="ScreenSpacePosition">The position of the current pixel in screen space (usually don't need to set this).</param>
 		/// <param name="Out">A floating-point value between 0 and 1 (usually plug this into the Alpha input on the master node).</param>
@@ -49,7 +50,7 @@ namespace Jick
 half jick_dither( half a, half2 screenPos, bool jitter )
 {
 	half2 uv = screenPos.xy * _ScreenParams.xy;
-	if ( jitter ) uv += _FrameCount % 4;
+	if ( jitter ) uv += uint( _FrameCount ) % 4;
 	half DITHER_THRESHOLDS[16] = {1.0/17.0,9.0/17.0,3.0/17.0,11.0/17.0,13.0/17.0,5.0/17.0,15.0/17.0,7.0/17.0,4.0/17.0,12.0/17.0,2.0/17.0,10.0/17.0,16.0/17.0,8.0/17.0,14.0/17.0,6.0/17.0};
 	return a - DITHER_THRESHOLDS[ ( ( uint( uv.x ) % 4 ) * 4 + uint( uv.y ) % 4 ) ];
 }
